requirements.txt
requests
pandas
streamlit
transformers
python-dotenv
